requirements.txt
# Engram - Unified Requirements

# Core Flask API
flask>=2.3.0
flask-cors>=4.0.0

# LLM and AI
openai>=1.0.0
vllm>=0.3.0

# Memory Management (Hippocampus)
letta

# Data Processing
python-dateutil>=2.8.0
numpy
requests>=2.31.0
